+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added new `ex.Camera.setStrategies()` and `ex.Camera.strategies` for additional control of strategy order
 - Fixed ex.Font measureText always using 10px sans-serif on the first call on some browsers
 - Added a new `ex.Sound({...})` option back constructor to set all the same props available on sound
 - Added a new `ex.SoundManager` type for managing groups of audio/sound effects/music volume in an easier way
@@ -125,6 +126,7 @@ const query = new Query({})
 
 ### Changed
 
+- Updated `ex.Camera.addStrategy()` to accept multiple strategies
 - Changed the behavior of `fromSpriteSheet(...)`, `fromSpriteSheetCoordinates({...})` and `clone()` of `ex.Animation` to return the subclass if called from there
 - Optimized BoundingBox.rayCast and BoundingBox.rayCastTime
 - Optimized BoundingBox.intersect(otherBoundingBox)